+++ xlive/Blam/Cache/TagGroups/render_model_definition.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include "Blam\Cache\DataTypes\BlamDataTypes.h"
-#include "Blam\Cache\DataTypes\Strings.h"
 #include "Blam\Cache\TagGroups.hpp"
 #include "Blam\Math\BlamMath.h"
