+++ xlive/Blam/Engine/cseries/cseries_strings.h
