+++ xlive/H2MOD/Modules/PlayerRepresentation/PlayerRepresentation.h
@@ -1,5 +1,5 @@
 #include "Blam\Cache\DataTypes\BlamPrimitiveType.h"
-#include "Blam\Cache\DataTypes\StringID.h"
+#include "Blam/Engine/tag_files/string_ids.h"
 #include "Blam\Cache\TagGroups\globals_definition.hpp"
 #include "Blam/Engine/game/players.h"
